update team to have text cond selector
