+++ libs/gi/page-team/src/CharacterDisplay/OptimizeChrome/optimizeChromeTheme.ts
@@ -18,43 +18,6 @@ export function elementHeaderGradientSx(elementKey: ElementKey) {
   }
 }
 
-export function teamSlotGradientSx(
-  elements: Array<ElementKey | undefined>,
-  selectedIndex: number
-) {
-  return (theme: Theme) => {
-    const backrgba = colorToRgbaString(
-      hexToColor(theme.palette.contentLight.main as string)!,
-      0.5
-    )!
-    const rgbas = [
-      backrgba,
-      ...elements.map((ele, i) => {
-        if (!ele) return 'rgba(0,0,0,0)'
-        const hex = theme.palette[ele].main as string
-        const color = hexToColor(hex)
-        if (!color) return 'rgba(0,0,0,0)'
-        return colorToRgbaString(color, selectedIndex === i ? 0.5 : 0.15)
-      }),
-    ]
-    const selectedEle = elements[selectedIndex]
-    const selectedRgb =
-      selectedEle && hexToColor(theme.palette[selectedEle].main as string)
-    const borderRgba =
-      (selectedRgb && colorToRgbaString(selectedRgb, 0.45)) ??
-      'rgba(200,200,200,0.45)'
-    return {
-      background: `linear-gradient(to right, ${rgbas
-        .map(
-          (rgba, i, arr) =>
-            `${rgba} ${i * (100 / arr.length) + 50 / arr.length}%`
-        )
-        .join(', ')})`,
-      borderBottom: `2px solid ${borderRgba}`,
-    }
-  }
-}
-
 export function chromeJoinedButtonSx(index: number, total: number) {
   return {
     flex: '1 1 0',
